Create Day 2
hoping this isn't considered cheating...and i know how to do this in my head i just couldn't put it all into code. But to further explain: the initial deposit is $100. profit grows by 20% every year. so in the first year your total is $120. in the second year the total ($120) grows by another 20% to make the total $144. by the third year you add another 20% of $144 to make a total of $172.8. The answer for the code is supposed to show how may years it takes to get the threshold of $170. after a little youtube tutorial for how to fine the time for exponential growth, i found that it takes 2.7 years to break the threshold of 170. add a math.ceil to round up and you get 3 years. ill probably lose some sleep over this and come back later to find a better way to do it all in code!